integration-tests/sceptre-project/templates/valid_template_yaml.py
import yaml


def sceptre_handler(scepter_user_data):
    template = {
        "Resources": {
            "WaitConditionHandle": {
                "Type": "AWS::CloudFormation::WaitConditionHandle",
                "Properties": {},
            }
        }
    }
    return yaml.dump(template)
