pyproject.toml
[tool.poetry]
name = "sceptre"
version = "4.2.2"
packages = [{ include = "sceptre" }]
readme = "README.md"
homepage = "https://github.com/Sceptre/sceptre"
repository = "https://github.com/Sceptre/sceptre"
documentation = "https://docs.sceptre-project.org"
authors = ["Sceptre <sceptreorg@gmail.com>"]
description = "An AWS Cloud Provisioning Tool"
keywords = ["aws", "cloud", "devops", "infrastructure", "tools", "cli"]
license = "Apache-2.0"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Environment :: Console",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11"
]

[tool.poetry.scripts]
"sceptre" = "sceptre.cli:cli"

[tool.poetry.plugins."sceptre.hooks"]
"asg_scheduled_actions" = "sceptre.hooks.asg_scaling_processes:ASGScalingProcesses"
"cmd" = "sceptre.hooks.cmd:Cmd"

[tool.poetry.plugins."sceptre.resolvers"]
"environment_variable" = "sceptre.resolvers.environment_variable:EnvironmentVariable"
"file_contents" = "sceptre.resolvers.file_contents:FileContents"
"stack_output" = "sceptre.resolvers.stack_output:StackOutput"
"stack_output_external" = "sceptre.resolvers.stack_output:StackOutputExternal"
"no_value" = "sceptre.resolvers.no_value:NoValue"
"select" = "sceptre.resolvers.select:Select"
"stack_attr" = "sceptre.resolvers.stack_attr:StackAttr"
"sub" = "sceptre.resolvers.sub:Sub"
"split" = "sceptre.resolvers.split:Split"
"join" = "sceptre.resolvers.join:Join"

[tool.poetry.plugins."sceptre.template_handlers"]
"file" = "sceptre.template_handlers.file:File"
"s3" = "sceptre.template_handlers.s3:S3"
"http" = "sceptre.template_handlers.http:Http"

[tool.poetry.dependencies]
python = ">=3.7,<3.12"
boto3 = "^1.20.27"
click = ">=7.0,<9.0"
cfn-flip = "^1.2.3"
deepdiff = "^5.5"
deprecation = "^2.0"
jinja2 = "^3.0"
jsonschema = "~3.2"
networkx = "~2.6"
packaging = ">=16.8,<22.0"
pyyaml = "^6.0"
sceptre-cmd-resolver = "^2.0"
sceptre-file-resolver = "^1.0"
colorama = ">=0.2.5,<0.4.4"
troposphere = { version = "^4", optional = true }
sphinx = { version = ">=1.6.5,<=5.1.1", optional = true }
sphinx-click = { version = ">=2.0.1,<4.0.0", optional = true }
sphinx-rtd-theme = { version = "0.5.2", optional = true }
sphinx-autodoc-typehints = { version = "1.19.2", optional = true }
docutils = { version = "<0.17", optional = true } # temporary fix for sphinx-rtd-theme==0.5.2, it depends on docutils<0.17

[tool.poetry.extras]
troposphere = ["troposphere"]
docs = [
  "sphinx",
  "sphinx-click",
  "sphinx-rtd-theme",
  "sphinx-autodoc-typehints",
  "docutils"
]

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
behave = "^1.2.5"
freezegun = ">=0.3.12,<0.4.0"
pygments = "^2.2"
pytest = "^6.2"
pytest-cov = "^2.11.1"
pytest-sugar = "^0.9.4"
readme-renderer = "^24.0"
requests-mock = "^1.9.3"
tox = "^3.23.0"
troposphere = "^4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
